Moved slider to SliderDemo view
